+++ src/features/main/section-like-course.tsx
@@ -5,7 +5,7 @@ import Spacer from '@/src/shared/ui/Spacer'
 import GridCourse from '@/src/features/course/grid-course'
 import { CourseType } from '@/src/entities/course/type'
 import NoLikedCourse from '@/src/shared/ui/NoLikedCourse'
-import { useGetMyProfile } from '@/src/entities/user/query'
+import { useGetMyProfile } from '@/src/entities/user'
 import { useGetLikeCourses } from '@/src/entities/course/query'
 import { getLoginUrl } from '@/src/entities/login/api'
 import { useRouter } from 'next/navigation'
