+++ src/shared/ui/RegionCascader.tsx
@@ -4,7 +4,7 @@ import { Cascader } from 'antd'
 import { getSeoulData } from '@/src/entities/place/api'
 import { DefaultOptionType } from 'antd/es/cascader'
 import { useMemo } from 'react'
-import { UserLikeRegionType } from '@/src/entities/user/type'
+import { UserLikeRegionType } from '@/src/entities/user'
 
 interface CascaderProps {
   firstRegion: string
